Course Project Weekly Update 4 Commit 1
Tasks now read out from the database and can be scrolled through by the user. Is currently a TextView for testing purposes, but will likely be replaced by a RecyclerView.

Tasks can also be deleted using the button in the top right corner (which is a tentative proof of concept, this will be moved.)

Alongside that, there are some general quality of life improvements and fixes to the database input method.