Course Project Project Submission (Code) Commit 1
This is the final submission for my course project. Additional changes may be appended before the end of the semester, but what exists here is a complete and finished program.

Calendar View shows the user tasks and they assign, the days they are due by, their completion status, and their difficulty. They can be toggled for completion or deleted at will. There is a backend database for this system which uses an SQL implementation.

Concepts View shows the user links they have saved. Only complete links with HTTPs and the like are fully compatible as of this commit. Users can click a button to go to the saved link, copy the link to their clipboard, or remove links as they see fit. There is a backend database for this as well.